algorithms/medium/1679-max-number-of-k-sum-pairs/src/test/java/ru/saidgadjiev/leetcode/easy/_1679/SolutionTest.java
package ru.saidgadjiev.leetcode.easy._1679;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

class SolutionTest {

    @Test
    void case1() {
        Solution solution = new Solution();
        int i = solution.maxOperations(new int[]{1, 2, 3, 4}, 5);
        Assertions.assertEquals(2, i);
    }

    @Test
    void case2() {
        Solution solution = new Solution();
        int i = solution.maxOperations(new int[]{3, 1, 3, 4, 3}, 6);
        Assertions.assertEquals(1, i);
    }

    @Test
    void case3() {
        Solution solution = new Solution();
        int i = solution.maxOperations(new int[]{2, 5, 4, 4, 1, 3, 4, 4, 1, 4, 4, 1, 2, 1, 2, 2, 3, 2, 4, 2}, 3);
        Assertions.assertEquals(4, i);
    }
}